backend/requirements.txt
djangorestframework==3.15.1
django-cors-headers